requirements.txt
# STREAM Anti-Corruption Engine — Dependencies

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Database
psycopg[binary]>=3.1.0
psycopg_pool>=3.1.0
python-dotenv>=1.0.0

# Data & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
imbalanced-learn>=0.11.0
networkx>=3.0

# AI Agent
langchain-openai>=0.3.0
langgraph>=0.4.0
langchain-core>=0.3.0
